+++ src/dragengine/src/debug/deDebugBlockInfo.cpp
@@ -1,22 +1,25 @@
-/* 
- * Drag[en]gine Game Engine
+/*
+ * MIT License
  *
- * Copyright (C) 2020, Roland Plüss (roland@rptd.ch)
- * 
- * This program is free software; you can redistribute it and/or 
- * modify it under the terms of the GNU General Public License 
- * as published by the Free Software Foundation; either 
- * version 2 of the License, or (at your option) any later 
- * version.
+ * Copyright (C) 2024, DragonDreams GmbH (info@dragondreams.ch)
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 #include <stdio.h>
@@ -62,7 +65,7 @@ pSpacing( 10 )
 	pViewTitle.TakeOver( engine.GetCanvasManager()->CreateCanvasText() );
 	pViewTitle->SetFont( pFont );
 	pViewTitle->SetColor( decColor( 1.0f, 1.0f, 1.0f ) );
-	pViewTitle->SetFontSize( pFont->GetLineHeight() );
+	pViewTitle->SetFontSize( ( float )pFont->GetLineHeight() );
 	pViewTitle->SetText( "" );
 	pViewTitle->SetOrder( 1.0f );
 	pViewTitle->SetSize( pTextSize( pFont, "" ) );
@@ -195,7 +198,7 @@ const decColor &color, const decColor &bgColor ){
 		entry->viewLabel.TakeOver( canvasManager.CreateCanvasText() );
 		entry->viewLabel->SetFont( pFont );
 		entry->viewLabel->SetColor( color );
-		entry->viewLabel->SetFontSize( pFont->GetLineHeight() );
+		entry->viewLabel->SetFontSize( ( float )pFont->GetLineHeight() );
 		entry->viewLabel->SetText( label );
 		entry->viewLabel->SetOrder( 1.0f );
 		entry->viewLabel->SetSize( pTextSize( pFont, label ) );
@@ -204,7 +207,7 @@ const decColor &color, const decColor &bgColor ){
 		entry->viewText.TakeOver( canvasManager.CreateCanvasText() );
 		entry->viewText->SetFont( pFont );
 		entry->viewText->SetColor( color );
-		entry->viewText->SetFontSize( pFont->GetLineHeight() );
+		entry->viewText->SetFontSize( ( float )pFont->GetLineHeight() );
 		entry->viewText->SetText( text );
 		entry->viewText->SetOrder( 2.0f );
 		entry->viewText->SetSize( pTextSize( pFont, text ) );
@@ -321,7 +324,7 @@ decPoint deDebugBlockInfo::pTextSize( const deFont &font, const char *text ){
 		
 		if( character == '\n' ){
 			textHeight += font.GetLineHeight();
-			lineWidth = 0.0f;
+			lineWidth = 0;
 			
 		}else if( character >= 0 ){
 			const deFontGlyph &glyph = font.GetGlyph( character );
